+++ lib/elixir/test/elixir/inspect_test.exs
@@ -224,12 +224,12 @@ defmodule Inspect.OthersTest do
   end
 
   test :external_elixir_funs do
-    bin = inspect(function(Enum.map/2))
+    bin = inspect(&Enum.map/2)
     assert bin == "&Enum.map/2"
   end
 
   test :external_erlang_funs do
-    bin = inspect(function(:lists.map/2))
+    bin = inspect(&:lists.map/2)
     assert bin == "&:lists.map/2"
   end