+++ lib/elixir/test/elixir/kernel/quote_test.exs
@@ -304,13 +304,7 @@ defmodule Kernel.QuoteTest.ImportsHygieneTest do
 
   defmacrop get_bin_size_with_function do
     quote do
-      function(size/1).("hello")
-    end
-  end
-
-  defmacrop get_bin_size_with_kernel_function do
-    quote do
-      Kernel.function(size/1).("hello")
+      (&size/1).("hello")
     end
   end
 
@@ -319,7 +313,6 @@ defmodule Kernel.QuoteTest.ImportsHygieneTest do
     assert get_bin_size == 5
     assert get_bin_size_with_partial == 5
     assert get_bin_size_with_function == 5
-    assert get_bin_size_with_kernel_function == 5
   end
 
   defmacrop get_dict_size do
@@ -343,7 +336,6 @@ defmodule Kernel.QuoteTest.ImportsHygieneTest do
     assert get_bin_size == 5
     assert get_bin_size_with_partial == 5
     assert get_bin_size_with_function == 5
-    assert get_bin_size_with_kernel_function == 5
   end
 
   defmacrop with_size do