+++ lib/elixir/test/elixir/stream_test.exs
@@ -83,7 +83,7 @@ defmodule StreamTest do
   test :repeatedly do 
     stream = Stream.repeatedly(fn -> 1 end)
     assert Enum.take(stream, 5) == [1,1,1,1,1]
-    stream = Stream.repeatedly(function(:random.uniform/0))
+    stream = Stream.repeatedly(&:random.uniform/0)
     [r1,r2] = Enum.take(stream, 2)
     assert r1 != r2
   end